fix: matching the docs routes
separating unique network and evm based parachains
